Nengo-wide default dtype can be set in RC
This makes it possible to set a default NumPy dtype that will be
used in all the places that we previously used 'float64'.

In doing some limited testing, using 'float32' results in little
to no speed difference, but a large reduction in memory usage.
'float16' reduces memory usage even more, but runs into floating
point math errors in many unit tests.

+++ nengo/neurons.py
@@ -55,9 +55,9 @@ class NeuronType(object):
 
         Parameters
         ----------
-        max_rates : ndarray(dtype=float64)
+        max_rates : ndarray
             Maximum firing rates of neurons.
-        intercepts : ndarray(dtype=float64)
+        intercepts : ndarray
             X-intercepts of neurons.
         """
         J_max = 0
